feat: Add dynamic routing to month page
